+++ app/models/business_subscription/part_time.rb
@@ -1,12 +1,9 @@
 module BusinessSubscription
   class PartTime
-    def name
-      "part_time"
-    end
-
     def price_id
       BusinessSubscription.price_ids[:part_time_plan]
     end
+    alias_method :plan, :price_id
 
     def price
       99