+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ### April
 
+* April 22 - Add link to conversations in navigation bar - #384
+* April 21 - Note interested role types on cold message form - #383
+* April 18 - Automatically send Developer and Business welcome emails - #378
+* April 18 - Added banner to connect with Joe to pay hiring fee - #380
+* April 12 - Added messaging tips for businesses - #376
+* April 10 - Added "Invisiblize" button on developer profiles (for admins) - #370
 * April 7 - Migrate email sending from Mailgun to Postmark - #365
 * April 7 - Added `/admin/conversations` - #368